Standardize public interfaces in central location
